+++ test/snapshot/snapshot.test.ts
@@ -1,0 +1,19 @@
+import { describe, expect, test } from 'bun:test'
+import { createRegistry } from '../../src/common/textmate/index.ts'
+import { getVSCodeTokens, renderSnapshot } from '../../src/snapshot/index.ts'
+import { read_testdata, TESTLANG_GRAMMARS } from '../testutil.ts'
+
+const SCOPE = 'source.xy'
+const registry = createRegistry(TESTLANG_GRAMMARS)
+
+describe('test snapshot testing', () => {
+	test('report OK', async () => {
+		const src = read_testdata('snap/snap.testlang')
+
+		const tokens = await getVSCodeTokens(registry, SCOPE, src)
+		const res = renderSnapshot(tokens, SCOPE)
+
+		const expected = read_testdata('snap/snap.testlang.snap')
+		expect(res).toEqual(expected)
+	})
+})